cut down main file and pulled objects and test data into seperate file for cleanliness
